+++ includes/SpellChecker.php
@@ -1,8 +1,8 @@
 <?php
 
 // turn on debug for localhost etc
-$whitelist = array('127.0.0.1', '::1', 'localhost');
-if (in_array($_SERVER['SERVER_NAME'], $whitelist)) {
+$whitelist = array('127.0.0.1', '::1', 'localhost', 'wealth-psychology.com', 'www.wealth-psychology.com');
+if (in_array($_SERVER['SERVER_NAME'], $whitelist) && is_admin()) {
     error_reporting(E_ALL);
     ini_set('display_errors', '1');
 }